End Screen, Replay and Menu Keys
Implemented End "Screen" with key listeners for replaying and going back to the menu respectively. Also included velocity modifier, so the modifier for the speed of the enemy ships is increased by .2 for every 10 points one scores. The base velocityModifier is 1, so if one scores 10 points, it goes up to 1.2. If 20 points, 1.4, and so on and so forth.